requirements.txt
# Instant Messenger — APC VUB 2025-2026
# Install with:  pip install -r requirements.txt

# Server
Flask==3.1.3
Flask-SocketIO==5.6.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.8

# Client (SocketIO connection from Tkinter)
# python-socketio is already listed above (shared library)

# Optional — only needed for inline image display in the chat
# If not installed, images will show as "[image]" placeholder
Pillow>=10.0.0
